Se sube el makefile corregido junto con la carpeta 2010 faltante

+++ Strength in Time/Descargarlosarchivosqueexisten.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+nclude <iostream>
 #include <fstream>
 #include <string>
 #include <cstdlib>
@@ -10,19 +10,24 @@ CURL* curl = curl_easy_init();
 if (curl) {
   // Establecer la URL de la solicitud HTTP GET
   curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+
   // Establecer que la solicitud debe ser de tipo HEAD
   curl_easy_setopt(curl, CURLOPT_NOBODY, 1L);
+
   // Ejecutar la solicitud
   CURLcode res = curl_easy_perform(curl);
+
   // Verificar si la solicitud se realizó correctamente
   if (res != CURLE_OK) {
     std::cerr << "Error al hacer la solicitud: " << curl_easy_strerror(res) << std::endl;
     curl_easy_cleanup(curl);
     return false;
   }
+
   // Obtener el código de respuesta HTTP
   long response_code;
   curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
+
   // Verificar si el archivo existe
   if (response_code == 200) {
     curl_easy_cleanup(curl);
@@ -36,7 +41,7 @@ if (curl) {
   return false;
  }
 }
-*/
+*/ 
 
 
 
@@ -52,20 +57,20 @@ int main(int argc,char **argv) {
     return 1;
   }
 
-  /* if (archivo_existe(url)) {
-          // Si el archivo existe, descargarlo
-	  */
-  while (std::getline(infile, url)) {
-    std::cout << "Descargando " << url << "..." << std::endl;
-
-    // Obtener el nombre del archivo a partir de la URL
-    std::string filename = url.substr(url.find_last_of("/") + 1);
-
-    // Descargar el archivo con wget
-    std::string command = "wget " + url + " -P " + folder_name + " -O " + folder_name + "/" + filename;
-    system(command.c_str());
+  if (archivo_existe(url)) {
+    // Si el archivo existe, descargarlo
+	  
+      while (std::getline(infile, url)) {
+	std::cout << "Descargando " << url << "..." << std::endl;
+    
+	// Obtener el nombre del archivo a partir de la URL
+	std::string filename = url.substr(url.find_last_of("/") + 1);
+    
+	// Descargar el archivo con wget
+	std::string command = "wget " + url + " -P " + folder_name + " -O " + folder_name + "/" + filename;
+	system(command.c_str());
+      }
   }
-
   infile.close();
 
   std::cout << "Descarga completada." << std::endl;
